# Cost-Aware Sort: Query Planner & Inverted-Index Sorter (2-200x faster)