--- a/adapters/handlers/rest/configure_api.go
+++ b/adapters/handlers/rest/configure_api.go
@@ -414,6 +414,7 @@ func MakeAppState(ctx context.Context, options *swag.CommandLineOptionsGroup) *s
 		TenantActivityWriteLogLevel: appState.ServerConfig.Config.TenantActivityWriteLogLevel,
 		QuerySlowLogEnabled:         appState.ServerConfig.Config.QuerySlowLogEnabled,
 		QuerySlowLogThreshold:       appState.ServerConfig.Config.QuerySlowLogThreshold,
+		InvertedSorterDisabled:      appState.ServerConfig.Config.InvertedSorterDisabled,
 	}, remoteIndexClient, appState.Cluster, remoteNodesClient, replicationClient, appState.Metrics, appState.MemWatch) // TODO client
 	if err != nil {
 		appState.Logger.
@@ -1723,6 +1724,7 @@ func initRuntimeOverrides(appState *state.State) {
 		registered.RevectorizeCheckDisabled = appState.ServerConfig.Config.RevectorizeCheckDisabled
 		registered.QuerySlowLogEnabled = appState.ServerConfig.Config.QuerySlowLogEnabled
 		registered.QuerySlowLogThreshold = appState.ServerConfig.Config.QuerySlowLogThreshold
+		registered.InvertedSorterDisabled = appState.ServerConfig.Config.InvertedSorterDisabled
 
 		cm, err := configRuntime.NewConfigManager(
 			appState.ServerConfig.Config.RuntimeOverrides.Path,