--- a/adapters/repos/db/index.go
+++ b/adapters/repos/db/index.go
@@ -59,7 +59,7 @@ import (
 	esync "github.com/weaviate/weaviate/entities/sync"
 	authzerrors "github.com/weaviate/weaviate/usecases/auth/authorization/errors"
 	"github.com/weaviate/weaviate/usecases/config"
-	runtimeConfig "github.com/weaviate/weaviate/usecases/config/runtime"
+	configRuntime "github.com/weaviate/weaviate/usecases/config/runtime"
 	"github.com/weaviate/weaviate/usecases/memwatch"
 	"github.com/weaviate/weaviate/usecases/modules"
 	"github.com/weaviate/weaviate/usecases/monitoring"
@@ -673,8 +673,9 @@ type IndexConfig struct {
 	LSMEnableSegmentsChecksumValidation bool
 	TrackVectorDimensions               bool
 	ShardLoadLimiter                    ShardLoadLimiter
-	QuerySlowLogEnabled                 *runtimeConfig.DynamicValue[bool]
-	QuerySlowLogThreshold               *runtimeConfig.DynamicValue[time.Duration]
+	QuerySlowLogEnabled                 *configRuntime.DynamicValue[bool]
+	QuerySlowLogThreshold               *configRuntime.DynamicValue[time.Duration]
+	InvertedSorterDisabled              *configRuntime.DynamicValue[bool]
 }
 
 func indexID(class schema.ClassName) string {