--- a/adapters/repos/db/inverted/analyzer.go
+++ b/adapters/repos/db/inverted/analyzer.go
@@ -17,6 +17,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/weaviate/weaviate/adapters/repos/db/helpers"
+	ent "github.com/weaviate/weaviate/entities/inverted"
 	"github.com/weaviate/weaviate/entities/models"
 )
 
@@ -99,7 +100,7 @@ func (a *Analyzer) TextArray(tokenization string, inArr []string) []Countable {
 // Int requires no analysis, so it's actually just a simple conversion to a
 // string-formatted byte slice of the int
 func (a *Analyzer) Int(in int64) ([]Countable, error) {
-	data, err := LexicographicallySortableInt64(in)
+	data, err := ent.LexicographicallySortableInt64(in)
 	if err != nil {
 		return nil, err
 	}
@@ -138,7 +139,7 @@ func (a *Analyzer) UUIDArray(in []uuid.UUID) ([]Countable, error) {
 func (a *Analyzer) IntArray(in []int64) ([]Countable, error) {
 	out := make([]Countable, len(in))
 	for i := range in {
-		data, err := LexicographicallySortableInt64(in[i])
+		data, err := ent.LexicographicallySortableInt64(in[i])
 		if err != nil {
 			return nil, err
 		}
@@ -151,7 +152,7 @@ func (a *Analyzer) IntArray(in []int64) ([]Countable, error) {
 // Float requires no analysis, so it's actually just a simple conversion to a
 // lexicographically sortable byte slice.
 func (a *Analyzer) Float(in float64) ([]Countable, error) {
-	data, err := LexicographicallySortableFloat64(in)
+	data, err := ent.LexicographicallySortableFloat64(in)
 	if err != nil {
 		return nil, err
 	}
@@ -168,7 +169,7 @@ func (a *Analyzer) Float(in float64) ([]Countable, error) {
 func (a *Analyzer) FloatArray(in []float64) ([]Countable, error) {
 	out := make([]Countable, len(in))
 	for i := range in {
-		data, err := LexicographicallySortableFloat64(in[i])
+		data, err := ent.LexicographicallySortableFloat64(in[i])
 		if err != nil {
 			return nil, err
 		}
@@ -214,7 +215,7 @@ func (a *Analyzer) Bool(in bool) ([]Countable, error) {
 // being an explicitly allowed value as well.
 func (a *Analyzer) RefCount(in models.MultipleRef) ([]Countable, error) {
 	length := uint64(len(in))
-	data, err := LexicographicallySortableUint64(length)
+	data, err := ent.LexicographicallySortableUint64(length)
 	if err != nil {
 		return nil, err
 	}