--- a/adapters/repos/db/inverted/searcher.go
+++ b/adapters/repos/db/inverted/searcher.go
@@ -79,23 +79,32 @@ func (s *Searcher) Objects(ctx context.Context, limit int,
 	filter *filters.LocalFilter, sort []filters.Sort, additional additional.Properties,
 	className schema.ClassName, properties []string,
 ) ([]*storobj.Object, error) {
+	beforeFilters := time.Now()
 	allowList, err := s.docIDs(ctx, filter, additional, className, limit)
 	if err != nil {
 		return nil, err
 	}
 	defer allowList.Close()
+	helpers.AnnotateSlowQueryLog(ctx, "build_allow_list_took", time.Since(beforeFilters))
+	helpers.AnnotateSlowQueryLog(ctx, "allow_list_doc_ids_count", allowList.Len())
 
 	var it docIDsIterator
 	if len(sort) > 0 {
+		beforeSort := time.Now()
 		docIDs, err := s.sort(ctx, limit, sort, allowList, className)
 		if err != nil {
 			return nil, fmt.Errorf("sort doc ids: %w", err)
 		}
+		helpers.AnnotateSlowQueryLog(ctx, "sort_doc_ids_took", time.Since(beforeSort))
 		it = newSliceDocIDsIterator(docIDs)
 	} else {
 		it = allowList.Iterator()
 	}
 
+	beforeObjects := time.Now()
+	defer func() {
+		helpers.AnnotateSlowQueryLog(ctx, "objects_by_doc_ids_took", time.Since(beforeObjects))
+	}()
 	return s.objectsByDocID(ctx, it, additional, limit, properties)
 }