--- a/adapters/repos/db/migrator.go
+++ b/adapters/repos/db/migrator.go
@@ -140,6 +140,7 @@ func (m *Migrator) AddClass(ctx context.Context, class *models.Class,
 			ShardLoadLimiter:                    m.db.shardLoadLimiter,
 			QuerySlowLogEnabled:                 m.db.config.QuerySlowLogEnabled,
 			QuerySlowLogThreshold:               m.db.config.QuerySlowLogThreshold,
+			InvertedSorterDisabled:              m.db.config.InvertedSorterDisabled,
 		},
 		shardState,
 		// no backward-compatibility check required, since newly added classes will