--- a/adapters/repos/db/repo.go
+++ b/adapters/repos/db/repo.go
@@ -230,6 +230,7 @@ type Config struct {
 	TenantActivityWriteLogLevel         *configRuntime.DynamicValue[string]
 	QuerySlowLogEnabled                 *configRuntime.DynamicValue[bool]
 	QuerySlowLogThreshold               *configRuntime.DynamicValue[time.Duration]
+	InvertedSorterDisabled              *configRuntime.DynamicValue[bool]
 }
 
 // GetIndex returns the index if it exists or nil if it doesn't