--- a/adapters/repos/db/sorter/inverted_sorter.go
+++ b/adapters/repos/db/sorter/inverted_sorter.go
@@ -162,7 +162,13 @@ func (is *invertedSorter) quantileKeysForDescSort(ctx context.Context, limit int
 ) [][]byte {
 	ob := is.store.Bucket(helpers.ObjectsBucketLSM)
 	totalCount := ob.CountAsync()
+	if totalCount == 0 {
+		// no objects, likely no disk segments yet, force a full index scan
+		return [][]byte{{0x00}}
+	}
 	matchRate := float64(ids.Len()) / float64(totalCount)
+	fmt.Printf("match rate: %.2f, total count: %d, ids len: %d\n",
+		matchRate, totalCount, ids.Len())
 	estimatedRowsHit := int(float64(limit) / matchRate * 2) // safety factor of 2
 	if estimatedRowsHit > totalCount {
 		helpers.AnnotateSlowQueryLogAppend(ctx, "sort_query_planner",
@@ -171,6 +177,9 @@ func (is *invertedSorter) quantileKeysForDescSort(ctx context.Context, limit int
 		// full scan, just return zero byte (effectively same as cursor.First())
 		return [][]byte{{0x00}}
 	}
+
+	fmt.Printf("estimated rows hit: %d, total count: %d, match rate: %.2f\n",
+		estimatedRowsHit, totalCount, matchRate)
 
 	neededQuantiles := totalCount / estimatedRowsHit
 	quantiles := invertedBucket.QuantileKeys(neededQuantiles)