… adapters/repos/db/inverted/delta_analyzer_test.go
… adapters/repos/db/inverted/filters_integration_test.go
… adapters/repos/db/inverted/objects_test.go
… adapters/repos/db/inverted/searcher_integration_test.go
… adapters/repos/db/sorter/inverted_sorter_no_race_test.go
… adapters/repos/db/sorter/inverted_sorter_race_test.go
… adapters/repos/db/sorter/inverted_sorter_test.go
… adapters/repos/db/sorter/lsm_sorter.go
… adapters/repos/db/sorter/query_planner_test.go
… entities/inverted/serialization.go
… entities/inverted/serialization_test.go
… test/helper/race_off.go
… test/helper/race_on.go
… tools/dev/config.runtime-overrides.yaml
… usecases/config/config_handler.go
… usecases/config/environment.go
… usecases/config/runtimeconfig.go