## TL;DR / Summary

> A new sorting implementation that provides between a 2x and 200x speed-up depending on the scenario.

Our existing sorting implementation needs to load every object matched by a filter. That makes it `O(n)` where `n` is the number of matches. This number can be very high. This PR adds a new inverted-indexed-based sorting method with a more variable cost. In the best-case scenario, the cost is just `O(n)` where `n` is the limit, which can be orders of magnitude lower than the number of matches. In the worst-case scenario, however, it becomes `O(n)` with `n` being the number of objects stored. To prevent this, a cost-based query planner is added to fall back to the objects-bucket strategy when the predicted cost of using the inverted index is too high.

It is fully backward and forward-compatible. No state is changed; it only uses a different read path based on indexes that already existed. This makes it very low risk.

## Performance Improvements 🚀 

| Filter     | Order   | Before (no page cache)   | Before (page cache)   | After (no page cache)   | After (page cache)   | Improvement (no page cache)   | Improvement (page cache)   |
|:-----------|:--------|:-------------------------|:----------------------|:------------------------|:---------------------|:------------------------------|:---------------------------|
| unfiltered | ASC     | 1023 ms                  | 170 ms                | 9 ms                    | 1 ms                 | **114x**                      | **170x**                   |
| unfiltered | DESC    | 1047 ms                  | 176 ms                | 8 ms                    | 1 ms                 | **131x**                      | **176x**                   |
| 75% match  | ASC     | 1676 ms                  | 418 ms                | 29 ms                   | 2 ms                 | **58x**                       | **209x**                   |
| 75% match  | DESC    | 1613 ms                  | 413 ms                | 31 ms                   | 2 ms                 | **52x**                       | **207x**                   |
| 10% match  | ASC     | 472 ms                   | 53 ms                 | 23 ms                   | 1 ms                 | **21x**                       | **53x**                    |
| 10% match  | DESC    | 475 ms                   | 51 ms                 | 24 ms                   | 2 ms                 | **20x**                       | **26x**                    |
| 5% match   | ASC     | 449 ms                   | 32 ms                 | 31 ms                   | 2 ms                 | **14x**                       | **16x**                    |
| 5% match   | DESC    | 361 ms                   | 28 ms                 | 27 ms                   | 3 ms                 | **13x**                       | **9x**                     |
| 1% match   | ASC     | 132 ms                   | 9 ms                  | 30 ms                   | 5 ms                 | **4x**                        | **2x**                     |
| 1% match   | DESC    | 149 ms                   | 10 ms                 | n/a                 | n/a               | **n/a**                      | **n/a**                   |


In the rows where values are `n/a` the query planner chose the old strategy, i.e. those scenarios are unchanged.

All evaluation was done locally on a 3-ish-year-old MacBook.

## How it works / Why is this so fast?

**Ascending Sorts**
This implementation uses the fact that an inverted index is sorted by a specific property's values. Imagine a filter that matches every single object in the database. In this case, all we'd have to do is start reading at the beginning and stop reading after `limit` elements. In practice this means, even if we had 100k matches, but a limit of 100, we can stop reading after 100 elements. In fact, this is what the new implementation does.

**Filter rates**
Now, imagine the same scenario as before, but rather than matching every element, the filter matches every other element. Let's pretend the filter does not correlate with the property in the sort clause. This means we now have a 50% match rate. Our expectation is that we need to now read around 200 elements to find 100 matches. That is twice as expensive as the unfiltered scenario, but still orders of magnitude better than scoring 100k matches.

**Query planning**
By now you may be wondering; but what if the filter matches just a single object and it happens to be the very last one. In that case, the existing strategy is likely much better. For this, a cost-based query planner is introduced that estimates the cost of both strategies and picks the cheaper one.

**Descending sorts**
The inverted index is sorted in ascending order, we cannot easily walk it backwards (without requiring an extreme amount of disk seeks). To overcome this, the implementation uses a simple trick. It estimates where matches are most likely to begin in the inverted index and reads it from there to the end. So—sticking with our example of 100k values and an expected match rate of 200 matches— we would want to start reading somewhere around element 99,800 and read to the end. If we find enough matches, we have narrowed the search space down dramatically. If not, we need to extend the search to the next window (99,600-99,800). The better our estimate the fewer windows (and therefore seeks), we will need. In practice we often just need 1 or 2 windows making DESC search almost as efficient as ASC search.

**Other Contributing factors**
* Walking an index bucket from left to right is always more efficient than random LSM access. It never needs to use the segment index, preventing many seeks, and allowing the OS to predict what page to read next
* The existing implementation has an additional penalty based on the fact that there is no dedicated "value index" for properties; it needs to read a full object store row. In the worst case, this means reading a big vector that's not needed at all. Then it needs to spend CPU cycles dealing with JSON. This penalty is less than that the algorithmic penalty described above, but it contributes as well.

## Where this implementation shines
* unfiltered search with a single sort clause
* filtered sort with a medium to high match rate
* Whenever there is a positive correlation between the filter and the sort parameter (e.g., imagine a database of cars and you set a filter for `brand=Ferrari` and sort by `horsepower DESC`. While Ferraris may be rare in the dataset, they are likely some of the highest horsepower cars, so those IDs would show up quickly. 

## Limitations and potential future considerations
* Only `text`, `int`, and `date` props are currently considered. This implementation _should_ work with any prop that has a roaring set inverted index bucket (including array types). However, I decided to stick with the most common types for the initial implementation to keep the scope down—especially for testing and validation. New data types can be added fairly easily if necessary.
* The query planner does not consider correlation. If a query is strongly negatively correlated with a sort parameter, the query planner will under-estimate the cost of the inverted index strategy. This could be overcome by introducing sampling to estimate correlation.
* For multiple search clauses the inverted strategy simply calls itself recursively. This is likely a bad choice unless you have a high number of duplicates in the primary clause. Ideally, the sub-query should have it's own query planning rather than being locked into a single strategy. I don't think there is a technical reason not to do this, I simply omitted this to keep the implementation simpler and easier to test. My assumption is that the majority of use cases need faster single-clause sorting, so this is an acceptable trade-off. That said, it could easily be fixed in the future if necessary.
* Overall, this is still a workaround—albeit a much better workaround than the initial one. In an ideal scenario we would have a column-based index per prop for _really_ efficient sorting. A new index types is a lot more work though and will not be backward-compatible. It's definitely more of a long-term improvement if it becomes necessary. For now, this implementation should already speed up many use cases that previously couldn't use sorting.

## Observability and Explainability

The existing slow log is extended with a human-readable explanation of how the query planner decides and what the actual cost was. It also give optimization hints (i.e. enabling an inverted index) if a prop is not indexed. 

<img width="1256" alt="image" src="https://github.com/user-attachments/assets/590c1811-3f23-4acd-bce4-10542fed0c05" />

No new Prometheus metrics were added as of now. A possible way to extend this would be with simple counters about which path was chosen as well as tracking the execution time of each strategy.

## Configuration

There is no configuration required. However, a simple flag `INVERTED_SORTER_DISABLED=<bool>` was added. It is mainly helpful for testing and forcing the query planner to ignore the new strategy. It also offers an additional safety net in case we missed a bug. Then it could quickly be turned off. For this purpose, the flag is also compatible with the new run-time override config feature and can be configured using the key `inverted_sorter_disabled: bool`.

## Testing

A very extensive test was added that builds a large test matrix of varying data types, sort plans, limits, match rates, and object counts. It only runs the full test set when the race detector is disabled. Otherwise, it would run forever, especially with weaker machines on CI. However, a much simpler test plan is used when the race detector is on to catch any potential races. Our CI pipeline already runs all test with and without the race detector (as part of the integration tests which re-run unit tests), so both scenarios are covered on CI.

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.

<!-- Uncomment the following section if this PR requires changes in related projects (e.g., documentation, client libraries).

GitHub actions will automatically create an issue in the corresponding repository for each checked box below. (See `.github/workflows/create-cross-functional-issues.yml`)

### Cross-functional impact

- [ ] This change requires public documentation (weaviate-io) to be updated. Check the box to automatically create a corresponding issue.
- Does it require a change in the client libraries? If yes, please check the boxes for the affected client libraries.
    - [ ] Python (weaviate-python-client)
    - [ ] JavaScript/TypeScript (typescript-client)
    - [ ] Go (weaviate-go-client)
    - [ ] Java (java-client)

-->